lasieve4e: cast printf "%d" arg to int
gnfs-lasieve4e.c: In function ‘do_scheduling’:
gnfs-lasieve4e.c:3580:5: warning: format ‘%d’ expects type ‘int’, but argument 5 has type ‘long int’

arithmetic with size_t can result in 32 or 64 bits depending
on platform, so cast down.

+++ src/lasieve4/gnfs-lasieve4e.c
@@ -3577,7 +3577,7 @@ void do_scheduling(struct schedule_struct *sched, u32_t ns, u32_t ot, u32_t s)
             continue;
 /* report SCHED_PATHOLOGY to http://mersenneforum.org/showthread.php?t=11430 */
           fprintf(stderr,"\rSCHED_PATHOLOGY q0=%u k=%d excess=%d                      \n",
-		  (unsigned int)special_q, k, sched->schedule[ll+1][k]-(sched->schedule[0][k]+sched->alloc));
+		  (unsigned int)special_q, k, (int)(sched->schedule[ll+1][k]-(sched->schedule[0][k]+sched->alloc)));
           longjmp(termination_jb, SCHED_PATHOLOGY);
         }
     }
